[OAPIF provider] Split qgsoapifprovider.h.cpp into as many files as classes
No code change. Just better reflect the class organization
