+++ src/providers/wfs/oapif/qgsoapifcollection.cpp
@@ -20,7 +20,6 @@ using namespace nlohmann;
 #include "qgsoapifcollection.h"
 #include "moc_qgsoapifcollection.cpp"
 #include "qgsoapifutils.h"
-#include "qgsoapifprovider.h"
 
 #include <set>
 
@@ -118,7 +117,7 @@ bool QgsOapifCollection::deserialize( const json &j, const json &jCollections )
       if ( spatial.is_object() && spatial.contains( "bbox" ) )
       {
         QgsCoordinateReferenceSystem crs( QgsCoordinateReferenceSystem::fromOgcWmsCrs(
-          QgsOapifProvider::OAPIF_PROVIDER_DEFAULT_CRS
+          OAPIF_PROVIDER_DEFAULT_CRS
         ) );
         if ( spatial.contains( "crs" ) )
         {
@@ -200,7 +199,7 @@ bool QgsOapifCollection::deserialize( const json &j, const json &jCollections )
           mBbox.set( values[0], values[1], values[2], values[3] );
           QgsLayerMetadata::SpatialExtent spatialExtent;
           spatialExtent.extentCrs = QgsCoordinateReferenceSystem::fromOgcWmsCrs(
-            QgsOapifProvider::OAPIF_PROVIDER_DEFAULT_CRS
+            OAPIF_PROVIDER_DEFAULT_CRS
           );
           mLayerMetadata.setCrs( spatialExtent.extentCrs );
           metadataExtent.setSpatialExtents( QList<QgsLayerMetadata::SpatialExtent>() << spatialExtent );
@@ -367,7 +366,7 @@ bool QgsOapifCollection::deserialize( const json &j, const json &jCollections )
   if ( mCrsList.isEmpty() )
   {
     QgsCoordinateReferenceSystem crs = QgsCoordinateReferenceSystem::fromOgcWmsCrs(
-      QgsOapifProvider::OAPIF_PROVIDER_DEFAULT_CRS
+      OAPIF_PROVIDER_DEFAULT_CRS
     );
     mLayerMetadata.setCrs( QgsCoordinateReferenceSystem::fromOgcWmsCrs( crs.authid() ) );
     mCrsList.append( crs.authid() );
