+++ src/providers/wfs/oapif/qgsoapifitemsrequest.cpp
@@ -19,6 +19,7 @@ using namespace nlohmann;
 #include "qgslogger.h"
 #include "qgsoapifitemsrequest.h"
 #include "moc_qgsoapifitemsrequest.cpp"
+#include "qgsoapifshareddata.h"
 #include "qgsoapifutils.h"
 #include "qgsproviderregistry.h"
 #include "qgsvectordataprovider.h"
